requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
beautifulsoup4==4.13.4
boto3==1.38.40
botocore==1.38.40
certifi==2025.6.15
charset-normalizer==3.4.2
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.3.0
executing==2.2.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
jiter==0.10.0
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kafka-python==2.2.12
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.6
openai==1.88.0
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
playwright==1.52.0
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyzmq==27.0.0
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
